podman networking
Since 0.8.0 KIND uses custom networks with docker to leverage the
embedded DNS server and other features.

This provides the same functionality for podman, however, due to
the difference networking implementation of both (libnetwork vs CNI)
podman uses a different networking model for cluster.
It creates one network per cluster with the same lifecycle of the
cluster, i.e. the network is deleted when the cluster is deleted.

fix load balancer for podman

docker uses a custom DNS server listening in 127.0.0.1:53 for
custom networks, however, podman does not.
The loadbalancer docker image does not have the same "network
magic" we are using for the node images, thus we need to configure
a different resolver depending on the provider.

Signed-off-by: Antonio Ojea <aojea@redhat.com>