podman localhost IPv6 portmapping
portmap plugin  does not support to do portmapping on the ::1 address,
because it does not have the route_localnet workaround implemented
in the kernel for IPv4.

If we want to do portmapping to the localhost address for IPv6 with
podman, we don't use it for the kubeconfig because it will not work.
Instead, we return the direct parameters, because podman only runs
in Linux, that guarantees that pods are reachable from the host.